% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysum.R
\name{mysum}
\alias{mysum}
\title{My Sum}
\usage{
mysum(x, y)
}
\arguments{
\item{x}{numeric value}

\item{y}{numeric value}
}
\value{
sum of the arguments
}
\description{
My Sum
}
\author{
Marlene Marchena
}
